+++ frontend/src/app/layout/side-panel/pages/sources/sources.component.html
@@ -1,0 +1,14 @@
+<div class="sources-container d-flex mt-4 p-2">
+    <div class="row">
+        <div class="col-md-12">
+            <h2>
+                Sources
+            </h2>
+            <ul class="sources-list mt-3">
+                <li *ngFor="let item of sources">
+                    {{ item }}
+                </li>
+            </ul>
+        </div>
+    </div>
+</div>
