+++ frontend/src/app/layout/side-panel/pages/sources/sources.component.scss
@@ -1,0 +1,24 @@
+html,
+body,
+* {
+    cursor: default !important;
+}
+
+.sources-list {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+}
+
+.sources-list li {
+    background: #F0F0F0;
+    border: 1px solid #eee;
+    padding: 0.75rem 1rem;
+    margin-bottom: 0.5rem;
+    border-radius: 8px;
+    transition: background 0.3s ease;
+}
+
+.sources-list li:hover {
+    background: #BED6FF;
+}
