+++ frontend/src/app/layout/side-panel/pages/sources/sources.component.ts
@@ -1,0 +1,18 @@
+import { CommonModule } from '@angular/common';
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-sources',
+  imports: [CommonModule],
+  templateUrl: './sources.component.html',
+  styleUrl: './sources.component.scss'
+})
+export class SourcesComponent {
+  sources = [
+    'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+    'Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+    'Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris.',
+    'Duis aute irure dolor in reprehenderit in voluptate velit esse.',
+    'Excepteur sint occaecat cupidatat non proident.'
+  ];
+}
